- alterações gerais - alteração de migrations (refazer) - refatoração de adição e remoção de produto

+++ database/migrations/2017_10_17_201713_create_itens_table.php
@@ -16,7 +16,7 @@ class CreateItensTable extends Migration
         Schema::create('itens', function (Blueprint $table) {
             $table->increments('id');
             $table->integer('qtd');
-            $table->double('total')->nullable();
+	        $table->decimal('total', 8,2)->nullable();
             $table->timestamps();
 
             $table->unsignedInteger('order_id');
